Create electricity_monitor.ino
This is the Arduino code we used to simulate electricity fault monitoring in our ESCOM project.  We tested it using a power supply and an extension cord (not an actual utility pole). The code was run in the Arduino IDE and sent simulated data  such as electricity status, coordinates, and fault indicators to ThingsSpeak Cloud for real-time monitoring and later integration with our Flutter-firebase app.